refactor(api): split oversized organizations/base-crud controllers under 500-line guard
Guards CI failed check:runtime-complexity (controllerLines cap of 500)
on organizations.controller.ts (891 lines) and base-crud.controller.ts
(507 lines). Root-caused by extracting responsibilities rather than
suppressing the guard:

- base-crud.controller.ts: moved the local AggregatePaginateResult type
  to base-crud.types.ts and the resolveScopeId tenancy-pointer helper to
  base-crud-scope.util.ts.
- organizations.controller.ts: removed six duplicated sub-resource
  methods (findAllBrands, findAllIngredients, findAllVideos, findAllTags,
  findAllPosts, findAllActivities) that were shadowed dead code behind
  OrganizationsRelationshipsController's identical routes; the live
  implementations already lived on the relationships controller.
  canUserReadEntity (the cross-tenant IDOR fix) is untouched beyond a
  doc-comment pointing at the relocated verifyOrganizationAccess.
- Converted constructor-DI/@Body()/@Query()/@Param() DTO imports on both
  organizations controllers back to value imports per
  check:di-value-imports (emitDecoratorMetadata needs runtime references,
  not type-only ones).
- Updated organizations.controller.spec.ts and
  organizations-relationships.controller.spec.ts to match: dropped the
  stale findAllIngredients test on the now-removed controller method, and
  fixed/added coverage on the relationships controller for the live
  findAllPosts/findAllIngredients/findAllVideos/findAllTags/findAllActivities
  behavior.

Verified: check:di-value-imports (0 violations, 2619 files),
check-runtime-complexity --base f79903f20632b223b400a2ae9b91879d42258a80
(0 violations), check:architecture (all guards pass), scoped
type-check --filter=@genfeedai/api (clean), and 81 passing tests across
organizations/base-crud controller specs.
